Add product quality filters. Move train_predict functions out of utils.
